refactor: Moved all test files into test folder
