+++ package.json
@@ -43,9 +43,9 @@
     "build:cli": "microbundle --compress=false --target=node -f cjs -i ./src/cli/index.ts -o build/cli/index.js",
     "build:api": "microbundle --compress=false -i ./src/api/index.ts -o build/api/index.js --tsconfig ./tsconfig.api.json && cp ./src/api/tonomy.css ./build/api/tonomy.css",
     "build": "yarn run build:sdk && yarn run build:cli && yarn run build:api && yarn run build:util",
-    "test:unit": "jest --selectProjects \"Unit tests\"",
     "test:setup": "./test-integration/build-and-run-services.sh start",
     "test:setup-down": "./test-integration/build-and-run-services.sh stop",
+    "test:unit": "jest --selectProjects \"Unit tests\"",
     "test:integration": "jest --detectOpenHandles --selectProjects \"Integration tests\"",
     "test": "jest --detectOpenHandles",
     "cli": "node ./build/cli/index.cjs",
