+++ src/cli/bootstrap/settings.ts
@@ -4,8 +4,6 @@ import { LoggerLevel } from '../../sdk/util/settings';
 // cannot use NODE_ENV as it is always "production" on `yarn run build`
 const env = process.env.NODE_ENV || 'development';
 
-console.log(`NODE_ENV=${env}`);
-
 type ConfigType = {
     environment: string;
     blockchainUrl: string;
