+++ test/controllers/user.unit.test.ts
@@ -1,7 +1,7 @@
 import { jsStorageFactory } from '../../src/cli/bootstrap/jsstorage';
 import { JsKeyManager } from '../../src/sdk/storage/jsKeyManager';
 import { User } from '../../src/sdk/controllers/User';
-import { setTestSettings } from '../../test-integration/helpers/settings';
+import { setTestSettings } from '../helpers/settings';
 
 setTestSettings();
