+++ test/util/request.unit.test.ts
@@ -5,7 +5,7 @@ import { Issuer } from '@tonomy/did-jwt-vc';
 import { toDid } from '../../src/sdk/util/ssi/did-jwk';
 import { LoginRequest, LoginRequestPayload } from '../../src/sdk/util/request';
 import { PublicKey } from '@wharfkit/antelope';
-import { setTestSettings } from '../../test-integration/helpers/settings';
+import { setTestSettings } from '../helpers/settings';
 
 setTestSettings();
