+++ test/util/ssi/did.unit.test.ts
